# documentation/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service to store a high-integrity event register.",
        "version": "1.0.0",
        "title": "Register"
    },
    "host": "localhost:8005",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/item/{item_hash}/entries": {
            "get": {
                "description": "Get the list of entries that cite the item, in reverse chronological order",
                "operationId": "getItemEntries",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "item_hash",
                        "description": "The item's hash",
                        "required": true,
                        "type": "string",
                        "format": "sha256"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entry"
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/entries/republish": {
            "post": {
                "description": "Republish given list of entries to a given routing key",
                "operationId": "republishEntries",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Entries to be republished and routing key to republish to",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "entries": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    },
                                    "minLength": 1
                                },
                                "routing_key": {
                                    "type": "string"
                                }
                            },
                            "required": ["entries", "routing_key"]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "republished_entries": {
                                    "description": "Successfully republished entries",
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    }
                                },
                                "entries_not_found": {
                                    "description": "Entries not found to republish (entry number greater than max entry)",
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/record/{field_value}/entries": {
            "get": {
                "description": "Retrieve all entries with primary key \"field_value\"",
                "operationId": "getRecordFieldValueEntries",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "field_value",
                        "description": "The record's primary key",
                        "required": true,
                        "type": "string",
                        "format": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entry"
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/records/{field_name}/{field_value}": {
            "get": {
                "description": "Get all records with the specified value in the specified field",
                "operationId": "getRecordsNameValue",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "field_name",
                        "description": "The name of the field to match against",
                        "required": true,
                        "type": "string",
                        "format": ""
                    },
                    {
                        "in": "path",
                        "name": "field_value",
                        "description": "The value of the field to query on",
                        "required": true,
                        "type": "string",
                        "format": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/records"
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/proof/register/{proof_identifier}": {
            "get": {
                "description": "Get the integrity proof for the whole register",
                "operationId": "getRegisterProof",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "proof_identifier",
                        "description": "Which proof algorithm to use (merkle:sha-256)",
                        "required": true,
                        "type": "string",
                        "format": "proof"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/register-proof"
                        }
                    }
                }
            }
        },
        "/proof/entry/{entry_number}/{total_entries}/{proof_identifier}": {
            "get": {
                "description": "Get the integrity proof for the specified entry",
                "operationId": "getEntryProof",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "entry_number",
                        "description": "Which entry to get the proof for",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "total_entries",
                        "description": "Number of entries in the register",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "proof_identifier",
                        "description": "merkle:sha-256",
                        "required": true,
                        "type": "string",
                        "format": "proof"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entry-proof"
                        }
                    }
                }
            }
        },
        "/proof/consistency/{total_entries_1}/{total_entries_2}/{proof_identifier}": {
            "get": {
                "description": "Get the consistency proof for the specified sub-trees",
                "operationId": "getConsistencyProof",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "total_entries_1",
                        "description": "The size of the smaller subtree",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "total_entries_2",
                        "description": "The size of the larger subtree",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "proof_identifier",
                        "description": "merkle:sha-256",
                        "required": true,
                        "type": "string",
                        "format": "proof"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/consistency-proof"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entry": {
            "type": "object",
            "properties": {
                "entry_number": {
                    "type": "integer"
                },
                "entry_timestamp": {
                    "type": "string"
                },
                "item_hash": {
                    "type": "string"
                },
                "key": {
                    "description": "Primary identifier for the register",
                    "type": "string"
                },
                "item_signature": {
                    "type": "string"
                }
            }
        },
...
        "infoset": {
            "type": "object",
            "properties": {
                "entry_number": {
                    "type": "integer"
                },
                "entry_timestamp": {
                    "type": "string"
                },
                "item_hash": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "item": {
                    "type": "object",
                    "description": "The schema of the item is variable and enforced by Validation API"
                }
            }
        },
        "records": {
            "type": "object",
            "description": "Note: contains object for each infoset in the register.",
            "properties": {
                "<primary_key>": {
                    "$ref": "#/definitions/infoset"
                },
                "<primary_key_2>": {
                    "$ref": "#/definitions/infoset"
                }
            }
        },
...
        "register-proof": {
            "type": "object",
            "properties": {
                "tree-head-signature": {
                    "type": "string",
                    "description": "(Pending)"
                },
                "proof-identifier": {
                    "type": "string"
                },
                "root-hash": {
                    "type": "string",
                    "description": "Root hash of the Merkle tree"
                },
                "timestamp": {
                    "type": "string",
                    "format": "datetime",
                    "description": "When the MTH was calculated"
                },
                "tree-size": {
                    "type": "integer",
                    "description": "Number of leafs in the Merkle Tree"
                }
            }
        },
        "entry-proof": {
            "type": "object",
            "properties": {
                "proof-identifier": {
                    "type": "string",
                    "description": "merkle:sha-256"
                },
                "entry-number": {
                    "type": "string",
                    "description": "The entry the audit proof is for"
                },
                "merkle-audit-path": {
                    "type": "array",
                    "description": "The audit path for the requested entry",
                    "items": {
                        "type": "string",
                        "format": "sha-256"
                    }
                }
            }
        },
        "consistency-proof": {
            "type": "object",
            "properties": {
                "proof-identifier": {
                    "type": "string"
                },
                "merkle-consistency-nodes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "description": "The proof of consistency between the trees"
                }
            }
        }
    }
}